text-to-speech/requirements.txt
Flask==3.0.3
azure-cognitiveservices-speech==1.37.0
python-dotenv==1.0.1
werkzeug==3.0.1
